完善C++版本TIC-80 API和精灵生成
- 补全Tic80.hpp缺失的wrapper函数和常量定义
- 实现Assets.hpp精灵/瓦片运行时生成逻辑
- 消除编译警告

+++ cpp/src/gameplay/Scene/CollisionSystem.hpp
@@ -91,7 +91,7 @@ public:
         const auto playerID = core::Context::get().getPlayerID();
         core::EntityManager::get().getPosition(playerID, &playerX, &playerY, &playerZ);
 
-        const auto& entity = entities[entityIndex];
+        const auto& entity = entities[static_cast<size_t>(entityIndex)];
         const float entityX = entity.entity.pos.x;
         const float entityY = entity.entity.pos.y;
